+++ src/_internal/pipe/index.ts
@@ -12,8 +12,12 @@ type FactoryProduct<
   K extends keyof F,
 > = (arg: ArgType<F>) => ArgType<Lookup<ExtractReturnTypes<T>, K, any>, any>
 
+type Awaited<T> = T extends PromiseLike<infer U> ? U : T
+
 export type ExtractReturnTypes<T extends Func1[]> = {
-  [P in keyof T]: T[P] extends (a: ArgType<T[P]>) => infer R ? R : never
+  [P in keyof T]: T[P] extends (a: ArgType<T[P]>) => infer R
+    ? Awaited<R>
+    : never
 }
 
 export type PipeFactoryResult<F extends SyncRules> = LaxReturnType<